requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mypy-extensions==0.4.3
openapi-codec==1.3.2
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
simplejson==3.17.6
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
